# Faster Q4_K on Metal