This PR improves token generation speed for `Q4_K` on Metal by a significant amount using ideas from PRs #2248, #2212 and #2188. The table gives token generation time in ms/t on M2 Max with 30-core GPU:

| Model | Master | This PR | Speedup |
|--:|--:|--:|--:|
| 7B | 23.7 | 19.5 | 21.5% |
|13B | 41.4 | 31.9 | 29.8% |
|33B | 99.4 | 73.6 | 35.0% |
|65B | 194.1 | 141.0 | 37.7% |