@ggerganov 2023-07-20
I can provide results from M1 Pro a bit later