@ggerganov 2023-07-20
| Model | Master | This PR |
|--:|--:|--:|
| 7B | 48.4 | 35.4 |
|13B | 49.4 | 63.2 |