@ikawrakow 2023-07-20
@ggerganov Any chance the 13B points are reversed? Else it would mean that on current master token prediction is about the same for 7B and 13B on your M1 Pro.